Enum implicit_clone::sync::IMap

pub enum IMap<K: Eq + Hash + ImplicitClone + 'static, V: PartialEq + ImplicitClone + 'static> {
    Static(&'static [(K, V)]),
    Rc(Arc<IndexMap<K, V>>),
}

An immutable hash map type inspired by Immutable.js.

This type is cheap to clone and thus implements ImplicitClone. It can be created based on a &'static [(K, V)], or based on a reference counted IndexMap.

This type has the least stable API at the moment and is subject to change a lot before the 1.0 release.

Variants

Static(&'static [(K, V)])

A (small) static map.

Rc(Arc<IndexMap<K, V>>)

An reference counted map.

Implementations

impl<K: Eq + Hash + ImplicitClone + 'static, V: PartialEq + ImplicitClone + 'static> IMap<K, V>

pub fn iter(&self) -> IMapIter<'_, K, V>

Return an iterator over the key-value pairs of the map, in their order.

pub fn keys(&self) -> IMapKeys<'_, K, V>

Return an iterator over the keys of the map, in their order.

pub fn values(&self) -> IMapValues<'_, K, V>

Return an iterator over the values of the map, in their order.

pub fn len(&self) -> usize

Return the number of key-value pairs in the map.

Computes in O(1) time.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Computes in O(1) time.

pub fn get<Q>(&self, key: &Q) -> Option<V>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Return a clone to the value stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(K, V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Return clones to the key-value pair stored for key, if it is present, else None.

Computes in O(1) time (average).

pub fn get_full<Q>(&self, key: &Q) -> Option<(usize, K, V)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Return item index, key and value

pub fn get_index(&self, index: usize) -> Option<(K, V)>

Get a key-value pair by index.

Valid indices are 0 <= index < self.len()

Computes in O(1) time.

pub fn get_index_of<Q>(&self, key: &Q) -> Option<usize>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Return item index, if it exists in the map.

Computes in O(1) time (average).

pub fn contains_key<Q>(&self, key: &Q) -> bool
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Return true if an equivalent to key exists in the map.

Computes in O(1) time (average).

pub fn last(&self) -> Option<(K, V)>

Get the last key-value pair.

Computes in O(1) time.